Add JSONRepository

+++ Models/Question.cs
@@ -2,13 +2,13 @@
 {
     public class Question
     {
-        public Question(int id, string title, string[] answers, string rightAnswer) {
+        public Question(long id, string title, string[] answers, string rightAnswer) {
             Id = id;
             Title = title;
             Answers = answers;
             RightAnswer = rightAnswer;
         }
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string[] Answers { get; set; }
         public string RightAnswer { get; set; }
